Added working functionality to send WNT to the borrower. I also refactored all components to funtions
